+++ packages/respect-core/src/utils/binary-content-type-checker.ts
